refactor(core): simplify model json parsing
