+++ packages/core/package.json
@@ -90,7 +90,7 @@
     "dotenv": "^16.4.5",
     "fetch-socks": "^1.3.0",
     "js-yaml": "4.1.0",
-    "jsonrepair": "3.12.0",
+    "jsonrepair": "3.14.0",
     "mdast-util-from-markdown": "^2.0.2",
     "mdast-util-to-markdown": "^2.1.2",
     "openai": "6.3.0",
