+++ packages/core/src/ai-model/prompt/extraction.ts
@@ -1,6 +1,6 @@
 import type { AIDataExtractionResponse, ServiceExtractParam } from '@/types';
 import { getPreferredLanguage } from '@midscene/shared/env';
-import { safeParseJson } from '../service-caller/json';
+import { parseModelResponseJson } from '../service-caller/json';
 import { extractXMLTag } from './util';
 
 export function buildTypeQueryDemandValue(
@@ -41,7 +41,10 @@ export function parseXMLExtractionResponse<T>(
 
   let data: T;
   try {
-    data = safeParseJson(dataJsonStr) as T;
+    data = parseModelResponseJson(dataJsonStr, {
+      source: 'generic-object',
+      requireObject: false,
+    }) as T;
   } catch (e) {
     throw new Error(`Failed to parse data-json: ${e}`);
   }
@@ -50,7 +53,10 @@ export function parseXMLExtractionResponse<T>(
   let errors: string[] | undefined;
   if (errorsStr) {
     try {
-      const parsedErrors = safeParseJson(errorsStr);
+      const parsedErrors = parseModelResponseJson(errorsStr, {
+        source: 'generic-object',
+        requireObject: false,
+      });
       if (Array.isArray(parsedErrors)) {
         errors = parsedErrors;
       }
