+++ packages/core/src/ai-model/service-caller/index.ts
@@ -56,8 +56,7 @@ import {
 } from './request-timeout';
 export {
   extractJSONFromCodeBlock,
-  normalJsonParser,
-  safeParseJson,
+  parseModelResponseJson,
 } from './json';
 export type { JsonParser } from './json';
